[Roles] Add admin and editor roles
